initial work on gluster multiple volume support.
Cleanup to the GlusterVolume class for multiple volumes

Set a default block size if not already defined.

Ensure that the default "working" directory is on the default file system.

Remove unused code from GlusterFileSystem

move and remove a single use method.

remove the uri to mount method as it was only used once

fixed boolean during check ouf same scheme

absolute URIs missing the schemas should be qualified against the file system, and not the default/working directory.